Geisha/discord_bot/requirements.txt
discord.py
openai
python-dotenv
watchdog